chore: remove the `syscalls` crate
It is difficult to debug an unexpected panic on QEMU when an user
process uses the same panic handler as the kernel. Also, there is no
preparation for implementing the microkernel. For example, process
blocking, which is necessary for message passing, is not implemented.
For these reasons, it is better not to switch to the user privilege.