chore: remove the process manager process
To remove the initialization in the user mode. Initialization in the
user mode needs the system calls. Eventually all system calls will block
the caller process to send a message, so it will be unable to call a
system call during the initialization because the caller process does
not exist. Therefore, all initialization must be done in the kernel
mode.