refactor: depend on cargo to build the kernel library and bootx64.efi
It is difficult to trace changes of files of source codes because these
crates depend on other crates in the ramen directory, and these
dependencies are often forgotten to write in the Makefile. Cargo exits
immediately if no changes occurred in the source codes. This commit
almost does not affect the building time.