refactor: create the `predefined_mmap` crate (#1020)
This commit moves the `constant` module in the `common` crate to the newly created `predefined_mmap` crate. The name "common" originates from "things which both the kernel and the bootloader codes need". However, the term is ambiguous, and I threw lots of codes into the crate. So this commit splits the `common` crate.