refactor(kernel): remove the `uefi` crate deps (#1022)
This commit removes the `uefi` dependency from the kernel to solve a problem caused by the workspace's property. The kernel has its global allocator, but the `uefi` crate also provides one. Although the kernel disables `uefi`'s one, it is not disabled because the `bootx64` crate enables it, so the compiler compiles the `uefi` crate with the allocation feature enabled. It causes global allocation conflicts. Removing the `uefi` crate dependency makes it possible to utilize Cargo's workspace, making the build speed faster.