refactor: set interrupt handlers written in Rust
Previously when an interrupt happens, an interrupt handler written in am assembly language was called. This handler pushed all registers, including callee-save registers, because I utilized these registers pushing to implement the process switch.  However, this was not a good way since this generates an unnecessary dependency between the interrupt code and the process implementation. The process switch should be able without the timer interrupts. Since this is the interrupt code, not all registers have to be saved. So I use the `x86-interrupt` calling convention and set the interrupt handlers with `set_handler_fn`.